Retrieve variables from datafiles.
Retrieves selected variables from the datafiles.
Will retrieve the identifier, name and number of dimensions.
Only valid variables will be presented as options.
Also updated the latest dump to have an unlimited size for the json
info and emptied the computation and datafile tables.

Don't forget to run `pip install Pydap` if not using **install.sh**